+++ app/javascript/debounce.js
@@ -1,0 +1,15 @@
+const debounce = (func, wait) => {
+  let timerId
+  return function (...args) {
+    if (timerId) {
+      clearTimeout(timerId)
+    }
+    timerId = setTimeout(() => {
+      func.apply(this, args)
+    }, wait)
+  }
+}
+
+export default {
+  debounce
+}
